websocket.js: Rework to take an 'endpoint' argument
Instead of hard-coding the endpoint, shuffle things around so the
factory only has a 'connect' method.  Calling connect (with an
'endpoint' argument) gives you a new object instance that holds the
wrapped-WebSocket state.  This means you can instantiate as many
websocket connections as you like:

  var websocket_a = $websocket.connect('ws://ws.a.net');
  var websocket_b = $websocket.connect('ws://ws.b.net');

and then emit and register on sockets independently.