websocket.js: Remove Underscore map/unique/flatten from dispatch.handler
Use the native Array.indexOf [1] to ensure unique callbacks.  The
'register' function already ensures unique callbacks for a particular
key, but with the 'handle' function's topic-prefix matching
(e.g. handlers for '/state' and '/state/get' will both be invoked for
a '/state/get' message) we may have duplicates due to matching several
keys.

Now that we're building an 'interested' array that is unique and flat,
we can replace the map/unique/flatten call with a native Array.forEach
[2].

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/indexOf
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/forEach