websocket.js: Remove Underscore extend from dispatch.handler
I'm not sure what this is doing.  It looks like 'extend(object)' is a
no-op [1].  Perhaps the original intent was to make a copy?  In that
case, perhaps the shallow-copy 'clone' would be better [2]?  In any
case, the benefits of a copy here (preferably deep) are that a series
of interested listener callbacks will get the original data regardless
of how previous callbacks may have altered the object they received.
In production, we only have a single functioning callback for each
topic, so we don't have to worry about this.  By dropping the _.extend
call, we get rid of our last remaining Underscore dependency.

If, in the future, we want to restore some kind of pre-callback copy,
there's a native shallow-copy example in Mozilla's forEach docs [3].
Extending that to perform deep copies of JSON-serialzable objects
shouldn't be hard, since objects and arrays are the only mutable
types you'd have to deal with.

[1]: http://underscorejs.org/#extend
[2]: http://underscorejs.org/#clone
[3]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/forEach#An_object_copy_function