websocket.js: Remove 'current' from 'register'
If wrapped_websocket.listeners[topic] was false (e.g. undefined), we
were creating a new array for current, but not assigning that array
back to wrapped_websocket.listeners[topic].  The new implementation
takes advantage of the fact that empty JavaScript arrays are truthy to
setup a new array when necessary, so we don't have to bother with an
additional variable.